requirements.txt
allure-pytest==2.8.32
allure-python-commons==2.8.32
apipkg==1.5
astroid==2.4.2
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
dacite==1.6.0
execnet==1.8.0
graphviz==0.16
idna==2.10
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
objgraph==3.5.0
packaging==20.8
pluggy==0.13.1
py==1.10.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
pytest-forked==1.3.0
pytest-rerunfailures==9.1.1
pytest-xdist==2.2.0
requests==2.25.1
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.3
visual-regression-tracker==4.7.0
webdriver-manager==3.3.0
wrapt==1.12.1
